feat: use rudder + segment in parallel
