+++ next.config.js
@@ -39,6 +39,10 @@ const nextConfig = {
   },
   rewrites: async () => [
     {
+      source: '/api/segment',
+      destination: 'https://api.segment.io/v1/batch',
+    },
+    {
       source: '/api/analytics/:path*',
       destination: 'https://near.dataplane.rudderstack.com/:path*',
     },
